Fix commit diff view for initial commits.
This always assumed that there was a diff available without checking
to make sure.  This change checks to ensure that a real diff is
possible, and if not, just takes every line from the b_blob and
marks it as an addition.

Thanks to Nate Haas (natehaas) for help with this patch.

Signed-off-by: Tim Lucas <t.lucas@toolmantim.com>