Append .local to the hostname if it's not already.
On Linux (at least on Ubuntu versions Karmic & Jaunty), the hostname of
the other bananajours cannot be resolved because the name resolution
doesn't search .local (MDNS).  I've tried a number of incantations to
Avahi configuration, /etc/resolv.conf and /etc/nsswitch.conf to get this
to happen but without success.

So, in order to get this to happen, we need to make sure all of the
published URIs use $hostname.local.

Also, all Bananajours will need to be updated, as this fix is on the
publishing side, not on the consumer side.