Another attempt to beat DNSSD::UnknownError
I refactored browser.rb so that all of DNSSD methods are executed
synchronously within one thread for browser.rb.  This makes it much
easier to catch errors because I don't need to worry what thread the
errors are occurring on - there can just be one handler within the browser
thread.